feat: enhance virtual event features with new components and settings
