+++ src/app/[id]/event/components/EventGallerySection.tsx
