+++ src/app/[id]/event/components/EventHeroSection.tsx
