+++ src/app/[id]/event/components/EventHostSection.tsx
