+++ src/app/[id]/event/components/EventLocation.tsx
