+++ src/app/[id]/event/components/TicketCard.tsx
