+++ src/app/[id]/event/components/VirtualEvent.tsx
